+++ lib/features/vaccination/exporter.dart
@@ -31,8 +31,8 @@ import 'package:flutter/material.dart';
 import 'package:csv/csv.dart';
 import 'package:solidpod/solidpod.dart';
 
+import 'package:healthpod/constants/vaccination_survey.dart';
 import 'package:healthpod/utils/normalise_timestamp.dart';
-import 'package:healthpod/constants/survey.dart';
 
 /// Handles exporting vaccination data from JSON files to a single CSV file.
 ///
