+++ lib/features/vaccination/importer.dart
@@ -32,7 +32,7 @@ import 'package:flutter/material.dart';
 import 'package:csv/csv.dart';
 import 'package:solidpod/solidpod.dart';
 
-import 'package:healthpod/constants/survey.dart';
+import 'package:healthpod/constants/vaccination_survey.dart';
 import 'package:healthpod/utils/format_timestamp_for_display.dart';
 import 'package:healthpod/utils/is_valid_timestamp.dart';
 import 'package:healthpod/utils/normalise_timestamp.dart';
